Add caching functionality to count_vowels function
This commit introduces a caching decorator to the count_vowels function. This cache will help improve performance when repeatedly calling this function with the same arguments, by storing the results and retrieving them from cache when needed. Functions to print cache info, cache parameters and clear the cache are also included.